(refactor): handler helper functions
Cleaned up the helper functions a bit. Added better parsing for
X-Forwarded-For headers as well as checking for X-Real-IP in  the
fetchIP function. Added a content-length header to the echo handler.
Modified fetchHeaders to provide all header values when more than one is
present by joining them with a comma. Headers is a map[string][]string,
which either requires us to ignore all values but the first, or to
display them.