(refactor): server package
Added a server package to refactor main to make the program more
readable. I added context to the application as well as the ability to
run as an HTTPS server if desired. The scaffolding is there for HTTPS,
but I do need to add CLI flags for TLS key and TLS cert files. I added a
graceful shutdown to the server using context and signal.Notify().